--- src/components/BotonComienzo.js
@@ -1,42 +1,0 @@
-import React, { useState, useEffect } from "react";
-import { FaArrowCircleUp } from "react-icons/fa";
-
-
-const BotonComienzo = () => {
-
-  const [showTopBtn, setShowTopBtn] = useState(false);
-
-  useEffect(() => {
-    window.addEventListener("scroll", () => {
-      if (window.scrollY > 400) {
-        setShowTopBtn(true);
-      }else {
-        setShowTopBtn(false);
-      }
-    });
-  }, []);
-
-  const goToTop = () => {
-    window.scrollTo({
-      top: 0,
-       behavior: "smooth",
-    });
-  };
-
-  return (
-    <div className="top-to-btm">
-
-      {" "}
-      {showTopBtn && (
-
-        <FaArrowCircleUp className="icon-position icon-style" onClick={goToTop} />
-
-      )}
-
-      {" "}
-    </div>
-  );
-};
-
-export default BotonComienzo;
-
