--- src/components/Curriculum.js
@@ -1,173 +1,0 @@
-import React from 'react'
-import "../style/curriculum.css";
-
-export const Curriculum = () => {
-  return (
-    <div className="pagina">
-      <br/>
-      <h1 className="titulo">Curriculum</h1>
-      <br />
-
-      <h2 className="cv">Experiencia</h2>
-      <br />
-      <h3><a href="https://www.vippinn.com" target="_blank" rel="noreferrer"><img src="/images/imgvippinn.png" className="imgcvv" alt='imagen'/></a> &nbsp;Vippinn</h3>
-        <p>Jornada parcial - <b>Actualidad</b>
-        <br />
-        Córdoba, Argentina</p>
-        <br />
-
-      <ul>
-        <li><b>Desarrolladora FrontEnd React Trainee - Pasante PIL </b>
-          <br />
-          &nbsp;<em>jun. 2023 - Actualidad</em>
-        </li>
-      </ul>
-      <br />
-      <br />
-
-      
-      <h3><a href="https://www.farmacity.com/" target="_blank" rel="noreferrer"><img src="/images/imgfty.png" className="imgcv" alt='imagen'/></a> &nbsp;Farmacity</h3>
-        <p>Jornada completa - <b>11 años 5 meses</b>
-        <br />
-        Córdoba, Argentina</p>
-        <br />
-
-      <ul>
-        <li><b>Asistente de Gerencia</b>
-          <br />
-          &nbsp;<em>sept. 2017 - may. 2023 · 5 años 9 meses</em>
-          <ul>
-            <br />
-            <li>Encargada en mi turno del local comercial y su administración</li>
-            <li>Equipo a cargo</li>
-            <li>Cierre de caja del personal y manejo de caja fuerte</li>
-            <li>Coordinación de horarios del personal</li>
-            <li>Distribución de tareas diarias</li>
-            <li>Trato con los proveedores</li>
-            <br />
-          </ul>
-        </li>
-        <li><b>Auxiliar de caja</b>
-          <br />
-          &nbsp;<em>jul. 2016 - ago. 2017 · 1 año 2 meses</em>
-          <ul>
-            <br />
-            <li>Cierre de caja del personal y manejo de caja fuerte</li>
-            <li>Distribución de tareas diarias</li>
-            <li>Equipo a cargo</li>
-            <br />
-          </ul>
-        </li>
-        <li><b>Cajero / Repositor</b>
-          <br />
-          &nbsp;<em>ene. 2012 - jun. 2016 · 4 años 6 meses</em>
-          <ul>
-            <br />
-            <li>Atención al cliente</li>
-            <li>Ventas</li>
-            <li>Manejo de caja</li>
-            <li>Reposición de mercadería</li>
-          </ul>
-        </li>
-      </ul>
-      <br />
-
-      <h3><a href="https://www.cinemarkhoyts.com.ar/" target="_blank" rel="noreferrer"><img src="/images/imghoyts.png" className="imgcv" alt='imagen'/></a> &nbsp;Cinemark Hoyts Argentina</h3>
-        <p>Jornada completa - <b>4 años 9 meses</b>
-        <br />
-        Córdoba, Argentina</p>
-        <br />
-
-      <ul>
-        <li><b>Oficial técnico</b>
-          <br />
-          &nbsp;<em>ago. 2007 - ene. 2012 · 4 años 6 meses</em>
-          <ul>
-            <br />
-            <li>Encargada de tareas de proyección de películas</li>
-            <li>Armado y desarmado de películas</li>
-            <li>Mantenimiento de los proyectores</li>
-            <li>Participación en tareas de mantenimiento del local</li>
-            <br />
-          </ul>
-        </li>
-        <li><b>Personal de ventas</b>
-          <br />
-          &nbsp;<em>may. 2007 - jul. 2007 · 3 meses</em>
-          <ul>
-            <br />
-            <li>Atención al cliente</li>
-            <li>Manejo de caja</li>
-            <li>Ventas</li>
-            <br />
-          </ul>
-        </li>
-      </ul>
-     
-      <br />
-      <hr />
-      <h2 className="cv">Cursos</h2>
-      <br />
-      <ul>
-      <li><img src="/images/imgispc.png" className="logoscv" alt='imagen'/>&nbsp;<b>Full Stack Python</b>
-          <br />
-          <em>ISPC - Instituto Superior Politécnico Córdoba - PIL 2023</em> - Actualidad</li>
-        <br />
-
-        <li><img src="/images/imgudemy.png" className="logoscv" alt='imagen'/>&nbsp;<b>Máster en React: Hooks, Context API, MERN Stack
-          ( MongoDB, ExpressJS, ReactJS y NodeJS ), JWT</b>
-          <br />
-          <em>Udemy</em> - Actualidad</li>
-        <br />
-
-        <li><img src="/images/imgseu.png" className="logoscv" alt='imagen'/>&nbsp;<b>Python: Primeros pasos</b>
-          <br />
-          <em>Secretaría de Extensión Universitaria UTN - FRC</em> - jul.2023</li>
-        <br />
-      
-        <li><img src="/images/imgseu.png" className="logoscv" alt='imagen'/>&nbsp;<b>Lógica y fundamentos de Programación</b>
-          <br />
-          <em>Secretaría de Extensión Universitaria UTN - FRC</em> - jun.2023</li>
-        <br />
-
-        <li><img src="/images/imgfcc.png" className="logoscv" alt='imagen'/>&nbsp;<b>Fundamentos de Git y GitHub</b>
-          <br />
-          <em>freeCodeCamp</em> - abr.2023</li>
-        <br />
-
-        <li><img src="/images/imgfcc.png"  className="logoscv" alt='imagen'/>&nbsp;<b>Algoritmos de JavaScript y proyectos de estructuras de datos</b>
-          <br />
-          <em>freeCodeCamp</em> - abr.2023</li>
-        <br />
-
-        <li><img src="/images/imgfcc.png"  className="logoscv" alt='imagen'/>&nbsp;<b>Fundamentos de React</b>
-          <br />
-          <em>freeCodeCamp</em> - ene.2023</li>
-        <br />
-
-        <li><img src="/images/imgfcc.png"  className="logoscv" alt='imagen'/>&nbsp;<b>Fundamentos de JavaScript</b>
-          <br />
-          <em>freeCodeCamp</em> - nov.2022</li>
-        <br />
-
-        <li><img src="/images/imgada.png" className="logoscv" alt='imagen'/>&nbsp;<b>Introducción al desarrollo FrontEnd</b>
-          <br />
-          <em>Ada ITW</em> - oct.2022</li>
-        <br />
-
-        <li><img src="/images/imgudemy.png" className="logoscv" alt='imagen'/>&nbsp;<b>Desarrollo web básico con HTML5, CSS3 y JavaScript</b>
-          <br />
-          <em>Udemy</em> - sept.2022</li>
-        <br />
-
-        <li><img src="/images/imgfcc.png" className="logoscv" alt='imagen'/>&nbsp;<b>Fundamentos de HTML5 y CSS3</b>
-          <br />
-          <em>freeCodeCamp</em> - jul.2022</li>
-      </ul>
-      <br />
-    </div>
-  )
-}
-
-
-
