--- src/components/Portafolio.js
@@ -1,16 +1,0 @@
-import React from 'react'
-import { ListadoTrabajos } from './ListadoTrabajos';
-//import { Habilidades } from './Habilidades';
-
-export const Portafolio = () => {
-  return (
-    
-    <div className="pagina">
-      <br/>
-      <h1 className="titulo">Portafolio</h1>
-
-      <ListadoTrabajos />
-
-    </div>
-  );
-};
