+++ src/components/Portafolio.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import { ListadoTrabajos } from './ListadoTrabajos';
+
+export const Portafolio = () => {
+    return (
+
+        <div className="pagina">
+            <br />
+            <h1 className="titulo">Portafolio</h1>
+
+            <ListadoTrabajos />
+
+        </div>
+    );
+};
