--- src/components/layout/Footer.js
@@ -1,17 +1,0 @@
-import React from 'react';
-//import { NavLink, Link } from 'react-router-dom';
-import "../../style/footer.css";
-
-
-export const Footer = () => {
-  return (
-    <footer className="footer">
-
-      Portafolio Escanes María Laura &copy; - 2023
-
-    </footer>
-  );
-};
-
-
-
