+++ src/index.css
@@ -1,83 +1,87 @@
 body {
-  background-color: #191919;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
-    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
-    sans-serif;
+    background-color: #191919;
+    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+        'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+        sans-serif;
 }
 
-*{
-  display: flexbox;
-  justify-content: center;
-  align-items: center;
+* {
+    display: flexbox;
+    justify-content: center;
+    align-items: center;
 }
 
-.layout{
-  width: 85%;
-  margin: 0 auto;
-}
- 
-.pagina{
-  color: white;
+.layout {
+    width: 85%;
+    margin: 0 auto;
 }
 
-.pagina .titulo{
-  font-size: 40px;
-  font-weight: bold;
-  color: #4150d0;
+.pagina {
+    color: white;
+}
+
+.pagina .titulo {
+    font-size: 40px;
+    font-weight: bold;
+    color: #4150d0;
 }
 
 /* Boton subir */
-.top-to-btm{
-  position: relative;
+.top-to-btm {
+    position: relative;
 }
 
-.icon-position{
-  position: fixed;
-  bottom: 40px;
-  right: 25px;
-  z-index: 20;
+.icon-position {
+    position: fixed;
+    bottom: 40px;
+    right: 25px;
+    z-index: 20;
 }
 
-.icon-style{
-  background-color: #fff;
-  border: 2px solid #fff;
-  border-radius: 50%;
-  height: 35px;
-  width: 35px;
-  color: #4150d0;
-  cursor: pointer;
-  animation: movebtn 3s ease-in-out infinite;
-  transition: all .5s ease-in-out;
+.icon-style {
+    background-color: #fff;
+    border: 2px solid #fff;
+    border-radius: 50%;
+    height: 35px;
+    width: 35px;
+    color: #4150d0;
+    cursor: pointer;
+    animation: movebtn 3s ease-in-out infinite;
+    transition: all .5s ease-in-out;
 }
 
-.icon-style:hover{
-  animation: none;
-  background: #4150d0;
-  color: #fff;
-  border: 2px solid #4150d0;
-  scale: 1.2;
+.icon-style:hover {
+    animation: none;
+    background: #4150d0;
+    color: #fff;
+    border: 2px solid #4150d0;
+    scale: 1.2;
 }
 
 @keyframes movebtn {
-  0%{
-    transform: translateY(0px);
-  }
-  25%{
-    transform: translateY(20px);
-  }
-  50%{
-    transform: translateY(0px);
-  }
-  75%{
-    transform: translateY(-20px);
-  }
-  100%{
-    transform: translateY(0px);
-  }
+    0% {
+        transform: translateY(0px);
+    }
+
+    25% {
+        transform: translateY(20px);
+    }
+
+    50% {
+        transform: translateY(0px);
+    }
+
+    75% {
+        transform: translateY(-20px);
+    }
+
+    100% {
+        transform: translateY(0px);
+    }
 }
 
 @media screen and (max-width: 460px) {
-  .pagina .titulo{
-   margin-top: 45px;    
-  }
+    .pagina .titulo {
+        margin-top: 45px;
+    }
 }
