--- src/jobs/trabajos.js
@@ -1,45 +1,0 @@
-export const trabajos = [
-  {
-    'id': 'Calculadora' ,
-    'nombre': 'Calculadora' ,
-    'tecnologias': 'HTML - CSS - JavaScript - React' ,
-    'url': 'https://calculadoraml.netlify.app/' , 
-    'codigo': 'https://github.com/EscanesMLaura/proyecto-calculadora.git' ,
-  },
-  {
-    'id': 'Lista-tareas' ,
-    'nombre': 'Lista de tareas' ,
-    'tecnologias': 'HTML - CSS - JavaScript - React' ,
-    'url': 'https://mis-tareasml.netlify.app/' ,
-    'codigo': 'https://github.com/EscanesMLaura/proyecto-mis-tareas.git' ,
-  },
-  {
-    'id': 'Deslizante-vertical' ,
-    'nombre': 'Deslizante vertical de imagenes' ,
-    'tecnologias': 'HTML - CSS - JavaScript' ,
-    'url': 'https://deslizante-verticalml.netlify.app/' ,
-    'codigo': 'https://github.com/EscanesMLaura/deslizante-vertical.git' ,
-  },
-  {
-    'id': 'Contador-agua',
-    'nombre': 'Contador agua' ,
-    'tecnologias': 'HTML - CSS - JavaScript' ,
-    'url': 'https://contador-aguaml.netlify.app/' ,
-    'codigo': 'https://github.com/EscanesMLaura/contador-agua.git' ,
-  },
-  {
-    'id': 'Desplazamiento-animacion' ,
-    'nombre': 'Desplazamiento animado' ,
-    'tecnologias': 'HTML - CSS - JavaScript' ,
-    'url': 'https://desplazamiento-animacionml.netlify.app/' ,
-    'codigo': 'https://github.com/EscanesMLaura/desplazamiento-animacion.git' ,
-  },
-  {
-    'id': 'Carrusel-imagenes' ,
-    'nombre': 'Carrusel de imagenes' ,
-    'tecnologias': 'HTML - CSS - JavaScript' ,
-    'url': 'https://carrusel-imagenesml.netlify.app/' ,
-    'codigo': 'https://github.com/EscanesMLaura/carrusel-imagenes.git' ,
-  },
-
-]
