--- src/style/curriculum.css
@@ -1,25 +1,0 @@
-.cv{
-  color: #c850c0;
-  }
-
-.logoscv{
-  width: 40px;
-  height: 40px;
-}
-
-.imgcv, .imgcvv{
-  width: 50px;
-  height: 50px;
-}
-
-.imgcvv:hover{
-  transition: all 300ms;
-  filter: drop-shadow(4px 4px 2px #c850c0);
-  scale: 1.2;
-}
-
-.imgcv:hover{
-  transition: all 300ms;
-  filter: drop-shadow(4px 4px 2px #4150d0);
-  scale: 1.2;
-}
