--- src/style/footer.css
@@ -1,16 +1,0 @@
-.footer{
-  line-height: 7px;
-  text-align: center;
-  font-size: 14px;
-  padding-top: 15px;
-  border-top: 2px solid white;
-  background: #191919;
-  color: white;
-  padding: 40px;
-}
-
-@media screen and (max-width: 670px) {
-  .footer{
-    width: 100%;
-  }
-}
