+++ src/style/headerNav.css
@@ -1,61 +1,60 @@
- .header{
-  display: flexbox;
-  justify-content: center;
-  align-items: center;
-  list-style: none;
-  height: 70px;
-  background-color: #191919;
+.header {
+    display: flexbox;
+    justify-content: center;
+    align-items: center;
+    list-style: none;
+    height: 70px;
+    background-color: #191919;
 }
 
-.header nav{
-  flex-grow: 1;
+.header nav {
+    flex-grow: 1;
 }
 
-.header nav ul{
-  height: 80px;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  list-style: none;
-  /*border-bottom: 2px solid white;*/
-  margin: 0px;
-  margin-top: 0px;
+.header nav ul {
+    height: 80px;
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    list-style: none;
+    margin: 0px;
+    margin-top: 0px;
 }
 
-.header li{
-  text-transform: uppercase;
-  font-weight: bold;
-  font-size: 15px;
-  margin-left: 15px;
-  margin-right: 15px;
+.header li {
+    text-transform: uppercase;
+    font-weight: bold;
+    font-size: 15px;
+    margin-left: 15px;
+    margin-right: 15px;
 }
 
-.header li:hover{
-  scale: 1.1;
+.header li:hover {
+    scale: 1.1;
 }
 
-.header li a{
-  color: white;
-  text-decoration: none;
-  transition: all 300ms;
+.header li a {
+    color: white;
+    text-decoration: none;
+    transition: all 300ms;
 }
 
 .header .active,
-.header a:hover{
-  color: #4150d0;
+.header a:hover {
+    color: #4150d0;
 }
 
 
 @media screen and (max-width: 460px) {
-  .header nav ul{
-    height: 130px;
-    display: flexbox;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    text-align: center;
+    .header nav ul {
+        height: 130px;
+        display: flexbox;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        text-align: center;
 
-    gap: 6px;
-    margin-left: -43px;
-  }
+        gap: 6px;
+        margin-left: -43px;
+    }
 }
