+++ src/style/listadoTrabajo.css
@@ -1,183 +1,183 @@
-*,*:before, *:after {
-  margin: 0;
-  position: 0;
-  outline: 0;
-  box-sizing: border-box;
+*,
+*:before,
+*:after {
+    margin: 0;
+    position: 0;
+    outline: 0;
+    box-sizing: border-box;
 }
 
 .proyectos {
-  font-family: "Roboto", sans-serif;
-  align-items: center;
-  justify-content: center;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(25rem, 1fr));
-  grid-auto-rows: 35rem;
+    font-family: "Roboto", sans-serif;
+    align-items: center;
+    justify-content: center;
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(25rem, 1fr));
+    grid-auto-rows: 35rem;
 
-  margin-left: -46px;
+    margin-left: -46px;
 }
 
 .contenedor {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  width: 940px;
-  gap: 15px;
-  margin: 50px;
-  justify-items: center;
-  align-items: center;
-} 
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    width: 940px;
+    gap: 15px;
+    margin: 50px;
+    justify-items: center;
+    align-items: center;
+}
 
 .card {
-  height: 360px;
-  width: 350px;
-  overflow: hidden;
-  margin: auto;
-  padding: 10px 30px;
-  border-radius: 2%;
-  border: 1px solid #fff;
-  transition: all .4s ease-in-out;
-  box-shadow: 0px 3px 14px #ddd;
-  text-align: center;
-  background: linear-gradient(-135deg, #c850c0, #4150d0);
+    height: 360px;
+    width: 350px;
+    overflow: hidden;
+    margin: auto;
+    padding: 10px 30px;
+    border-radius: 2%;
+    border: 1px solid #fff;
+    transition: all .4s ease-in-out;
+    box-shadow: 0px 3px 14px #ddd;
+    text-align: center;
+    background: linear-gradient(-135deg, #c850c0, #4150d0);
 }
 
 .card:hover {
-  height: 550px;
-  width: 350px;
-  border-radius: 5px;
-  box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.5);
+    height: 550px;
+    width: 350px;
+    border-radius: 5px;
+    box-shadow: 0px 2px 10px rgba(0, 0, 0, 0.5);
 }
 
 .card-encabezado img {
-  position: relative;
-  border-radius: 10%;
-  display: block;
-  height: 300px;
-  width: 290px;
-  object-fit: cover;
-  margin: 15px auto;
-  transition: all 0.3s ease;
+    position: relative;
+    border-radius: 10%;
+    display: block;
+    height: 300px;
+    width: 290px;
+    object-fit: cover;
+    margin: 15px auto;
+    transition: all 0.3s ease;
 }
 
 .card-encabezado {
-  position: relative;
+    position: relative;
 }
 
 .card-encabezado:before,
 .card-encabezado:after {
-  content: " ";
-  position: absolute;
-  opacity: .7;
+    content: " ";
+    position: absolute;
+    opacity: .7;
 }
 
 .card-iconos {
-  display: flex;
-  justify-content: center;
-  position: relative;
-  margin-bottom: 15px;
-  list-style: none;
+    display: flex;
+    justify-content: center;
+    position: relative;
+    margin-bottom: 15px;
+    list-style: none;
 }
 
 .card-iconos li {
-  border-radius: 50%;
-  background-color: #fff;
-  height: 40px;
-  width: 40px;
-  margin: 0 5px;
-  line-height: 40px;
-  text-align: center;
-  transform: scale(0);
-  transition: all 0.5s ease;
+    border-radius: 50%;
+    background-color: #fff;
+    height: 40px;
+    width: 40px;
+    margin: 0 5px;
+    line-height: 40px;
+    text-align: center;
+    transform: scale(0);
+    transition: all 0.5s ease;
 }
 
 .desplazar {
-  opacity: 0;
-  transition: all .5s ease;
-  transform: translateY(-45px);
+    opacity: 0;
+    transition: all .5s ease;
+    transform: translateY(-45px);
 }
 
 .agrandar {
-  transform: scale(0);
-  opacity: 0;
-  transition: all .4s ease;
+    transform: scale(0);
+    opacity: 0;
+    transition: all .4s ease;
 }
 
 .card:hover .agrandar {
-  opacity: 1;
-  transform: scale(1);
+    opacity: 1;
+    transform: scale(1);
 }
 
 .card:hover .desplazar {
-  transform: translateY(0px);
-  opacity: 1;
+    transform: translateY(0px);
+    opacity: 1;
 }
 
-ion-icon{
-  color:  #191919;
-  width: 25px;
-  height: 25px;
-  padding: 7px;
+ion-icon {
+    color: #191919;
+    width: 25px;
+    height: 25px;
+    padding: 7px;
 }
 
-ion-icon:hover{
-  color: #c850c0;
+ion-icon:hover {
+    color: #c850c0;
 }
 
 .card-iconos li a {
-  height: 100%;
-  width: 100%;
-  border-radius: 50%;
-  background: #fff;
-  display: block;
-  line-height: 40px;
-  transition: all 0.3s;
+    height: 100%;
+    width: 100%;
+    border-radius: 50%;
+    background: #fff;
+    display: block;
+    line-height: 40px;
+    transition: all 0.3s;
 }
 
 .card-iconos li a:hover {
-  transform: scale(1.2);
-  box-shadow: 0 0 5px #7813d6;
+    transform: scale(1.2);
+    box-shadow: 0 0 5px #7813d6;
 }
 
 .card-iconos li:nth-child(1) {
-  transition-delay: 0.2s;
+    transition-delay: 0.2s;
 }
 
 .card-iconos li:nth-child(2) {
-  transition-delay: 0.4s;
+    transition-delay: 0.4s;
 }
 
 .desplazar:nth-child(1) {
-  transition-delay: 0.3s;
+    transition-delay: 0.3s;
 }
 
 .desplazar:nth-child(3) {
-  transition-delay: 0.6s;
+    transition-delay: 0.6s;
 }
 
 .info-personal:nth-child(1) {
-  font-size: 15px;
-  font-weight: bold;
-  padding: 10px;
-  line-height: 25px;
+    font-size: 15px;
+    font-weight: bold;
+    padding: 10px;
+    line-height: 25px;
 }
 
 .info-personal:nth-child(3) {
-  margin: 15px 0;
+    margin: 15px 0;
 }
 
 @keyframes fadeInUp {
-  from {
-    opacity: 0;
-    transform: translateY(20px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
+    from {
+        opacity: 0;
+        transform: translateY(20px);
+    }
+
+    to {
+        opacity: 1;
+        transform: translateY(0);
+    }
 }
 
 .animated-card {
-  animation: fadeInUp 1s ease-in-out;
+    animation: fadeInUp 1s ease-in-out;
 }
-
-
-
